backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2
django-cors-headers==4.3.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
cryptography==41.0.7
python-decouple==3.8
pillow==10.1.0
requests==2.31.0
django-ratelimit==4.1.0
drf-yasg==1.21.7
django-extensions==3.2.3

# Arabic processing
PyArabic==0.6.15
libqutrub==0.0.15

# Vector search
pgvector==0.2.4

# Testing
pytest==7.4.3
pytest-django==4.8.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code quality
black==23.12.1
flake8==6.1.0
isort==5.13.2
mypy==1.7.1

# Production
gunicorn==21.2.0
python-dotenv==1.0.0
